+++ CHANGELOG.md
@@ -1,14 +1,25 @@
 # Release Notes
 
-## [Unreleased](https://github.com/ixocreate/framework/compare/0.2.0...develop)
+## [Unreleased](https://github.com/ixocreate/framework/compare/0.3.0...develop)
 
-## [v0.2.0 (2019-04-23)](https://github.com/ixocreate/framework/compare/0.1.0...0.2.0)
+## [v0.3.0 (2019-05-06)](https://github.com/ixocreate/framework/compare/0.2.1...0.3.0)
+### Changed
+- Upgrade to Application v0.2: remove PublishDefinitionBootstrapItem
+- Upgrade to Admin Package v0.3 (via CMS, Media, Registry, Translation)
 
+## [v0.2.1 (2019-04-25)](https://github.com/ixocreate/framework/compare/0.2.0...0.2.1)
+### Changed
+- Upgrade to Media Package v0.2
+
+## [v0.2.0 (2019-04-23)](https://github.com/ixocreate/framework/compare/0.1.1...0.2.0)
 ### Changed
 - Upgrade to Admin Package v0.2
 - Remove command bus defaults configuration (moved to Command Bus Package)
 
-## [v0.1.0 (2019-04-19)](https://github.com/ixocreate/framework/compare/master...0.1.0)
+## [v0.1.1 (2019-04-20)](https://github.com/ixocreate/framework/compare/0.1.0...0.1.1)
+### Added
+- Asset Package Dependency
 
+## [v0.1.0 (2019-04-19)](https://github.com/ixocreate/framework/compare/master...0.1.0)
 ### Changed
 - Consolidate Package